docker/app/requirements.txt
flask==2.3.3
psycopg2-binary==2.9.7
redis==4.6.0